+++ tests/superinstructions/test.sh
@@ -19,6 +19,8 @@ mkdir -p $SCRIPT_PATH/results/
 ## extract expected results
 tar xf $SCRIPT_PATH/expected-results.tar.bz2 -C $SCRIPT_PATH/
 
+NEEDS_UPDATE=false
+
 function runBenchmark {
   BENCH=$1
   HARNESS="${SOM_DIR}/som -si -Dsi.output=${SCRIPT_PATH}/results/${BENCH}.txt \
@@ -27,6 +29,9 @@ function runBenchmark {
   $HARNESS $@
   
   diff ${SCRIPT_PATH}/expected-results/${BENCH}.txt ${SCRIPT_PATH}/results/${BENCH}.txt
+  if [ $? -ne 0 ]; then
+    NEEDS_UPDATE=true
+  fi
 }
 
 runBenchmark LanguageFeatures.Fibonacci    1 0 2
@@ -56,7 +61,7 @@ runBenchmark Storage     1 0 2
 runBenchmark Sieve       1 0 2
 runBenchmark Towers      1 0 2
 
-if [ "$1" = "update" ]
+if [ "$1" = "update" ] && [ "$NEEDS_UPDATE" = true ]
 then
   ## move old results out of the way, and new results to expected folder
   rm -Rf $SCRIPT_PATH/old-results
